+++ api/pkg/controller/user-cluster-controller-manager/nodecsrapprover/README.md
@@ -1,0 +1,5 @@
+# `nodescrapprover` controller
+
+Used by openshift to autoapprove CSRs created by nodes for serving certs.
+
+TODO: Fix https://github.com/kubermatic/kubermatic/issues/4301 and use this for Kubernetes as well.
